pyproject.toml
[project]
name = "mlops-traffic-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.42.9",
    "evidently==0.4.19",
    "fastapi>=0.124.4",
    "h2o>=3.46.0.9",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "mlflow>=3.7.0",
    "numpy<2",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn<1.4",
    "seaborn>=0.13.2",
    "sniffio>=1.3.1",
    "tabulate>=0.9.0",
    "uvicorn>=0.38.0",
    "xgboost>=3.1.2",
]
